Added basic command completion
In order to finish up the tab completion, a few commands will need to be
added in order to allow for context-dependent completion.

For example, when the user types `minecraft_server stop [TAB]`, they
should only be provided with worlds that are currently running.  The
same goes with enabling/disabling worlds, etc.