Overviewer settings to use env vars for paths
Currently hardcoded paths for overview-settings.py
make it hard to keep the configs in git(...) repo
as they force the server backup/map files to always
be in exact absolute path on the machine.

(While also it cannot be always re-generated
 if user wants to tweak the rendering settings or such.)

For this lets not hardcode absolute paths, only relative,
relative to the environment variables we already have available
in msctl, so we can export them and use them as env vars inside python
settings file.